# frontend/docs/openapi-target.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Land Valuation API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/ai-assistant/sessions": {
      "post": {
        "tags": [
          "ai-assistant"
        ],
        "summary": "Create Assistant Session",
        "operationId": "create_assistant_session_api_v1_ai_assistant_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssistantSessionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantSessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/ai-assistant/sessions/{session_id}": {
      "get": {
        "tags": [
          "ai-assistant"
        ],
        "summary": "Get Assistant Session",
        "operationId": "get_assistant_session_api_v1_ai_assistant_sessions__session_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantSessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai-assistant/sessions/{session_id}/progress": {
      "get": {
        "tags": [
          "ai-assistant"
        ],
        "summary": "Get Assistant Progress",
        "operationId": "get_assistant_progress_api_v1_ai_assistant_sessions__session_id__progress_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantProgressResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai-assistant/sessions/{session_id}/messages": {
      "post": {
        "tags": [
          "ai-assistant"
        ],
        "summary": "Send Assistant Message",
        "operationId": "send_assistant_message_api_v1_ai_assistant_sessions__session_id__messages_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssistantMessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantMessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai-assistant/sessions/{session_id}/questions": {
      "post": {
        "tags": [
          "ai-assistant"
        ],
        "summary": "Ask Assistant Question",
        "operationId": "ask_assistant_question_api_v1_ai_assistant_sessions__session_id__questions_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssistantQuestionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantQuestionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "AssistantSessionCreate": {
        "properties": {
          "case_id": {
            "type": "string",
            "format": "uuid",
            "title": "Case Id"
          },
          "form_instance_id": {
            "type": "string",
            "format": "uuid",
            "title": "Form Instance Id"
          },
          "selected_form_type": {
            "type": "string",
            "pattern": "^F03$",
            "title": "Selected Form Type",
            "default": "F03"
          }
        },
        "type": "object",
        "required": [
          "case_id",
          "form_instance_id"
        ],
        "title": "AssistantSessionCreate"
      },
      "AssistantSessionResponse": {
        "properties": {
          "assistant_session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Assistant Session Id"
          },
          "case_id": {
            "type": "string",
            "format": "uuid",
            "title": "Case Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "form_instance_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Form Instance Id"
          },
          "current_step": {
            "type": "string",
            "title": "Current Step"
          },
          "selected_form_type": {
            "type": "string",
            "title": "Selected Form Type"
          },
          "missing_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Fields"
          },
          "missing_documents": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Documents"
          },
          "last_tool_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Tool Name"
          },
          "last_tool_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Tool Status"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "prompt_version": {
            "type": "string",
            "title": "Prompt Version"
          },
          "session_status": {
            "type": "string",
            "title": "Session Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "assistant_session_id",
          "case_id",
          "user_id",
          "form_instance_id",
          "current_step",
          "selected_form_type",
          "missing_fields",
          "missing_documents",
          "last_tool_name",
          "last_tool_status",
          "provider",
          "model_id",
          "prompt_version",
          "session_status",
          "created_at",
          "updated_at"
        ],
        "title": "AssistantSessionResponse"
      },
      "AssistantMessageRequest": {
        "properties": {
          "content": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Content"
          },
          "confirmed_fields": {
            "additionalProperties": true,
            "type": "object",
            "title": "Confirmed Fields"
          },
          "confirmed_candidate_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Confirmed Candidate Ids"
          },
          "confirm_apply": {
            "type": "boolean",
            "title": "Confirm Apply",
            "default": false
          },
          "run_calculation": {
            "type": "boolean",
            "title": "Run Calculation",
            "default": false
          },
          "run_validation": {
            "type": "boolean",
            "title": "Run Validation",
            "default": false
          },
          "generate_report_pdf": {
            "type": "boolean",
            "title": "Generate Report Pdf",
            "default": false
          },
          "nearest_facility": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NearestFacilityRequest"
              },
              {
                "type": "null"
              }
            ]
          },
          "confirm_action": {
            "type": "boolean",
            "title": "Confirm Action",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "AssistantMessageRequest"
      },
      "AssistantMessageResponse": {
        "properties": {
          "session": {
            "$ref": "#/components/schemas/AssistantSessionResponse"
          },
          "reply": {
            "type": "string",
            "title": "Reply"
          },
          "progress": {
            "$ref": "#/components/schemas/AssistantProgressResponse"
          },
          "tools": {
            "items": {
              "$ref": "#/components/schemas/ToolExecutionResponse"
            },
            "type": "array",
            "title": "Tools"
          }
        },
        "type": "object",
        "required": [
          "session",
          "reply",
          "progress",
          "tools"
        ],
        "title": "AssistantMessageResponse"
      },
      "AssistantQuestionRequest": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 2,
            "title": "Question"
          },
          "as_of_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of Date"
          },
          "document_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 10,
            "title": "Document Types"
          },
          "limit": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 5
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "question"
        ],
        "title": "AssistantQuestionRequest"
      },
      "AssistantQuestionResponse": {
        "properties": {
          "assistant_session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Assistant Session Id"
          },
          "answer_status": {
            "$ref": "#/components/schemas/KnowledgeAnswerStatus"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "generation_mode": {
            "type": "string",
            "title": "Generation Mode"
          },
          "next_action": {
            "type": "string",
            "title": "Next Action"
          },
          "clarification_question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clarification Question"
          },
          "claims": {
            "items": {
              "$ref": "#/components/schemas/AssistantClaim"
            },
            "type": "array",
            "title": "Claims"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/AssistantCitation"
            },
            "type": "array",
            "title": "Citations"
          },
          "unreadable_sources": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeUnreadableSource"
            },
            "type": "array",
            "title": "Unreadable Sources"
          }
        },
        "type": "object",
        "required": [
          "assistant_session_id",
          "answer_status",
          "answer",
          "generation_mode",
          "next_action"
        ],
        "title": "AssistantQuestionResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "NearestFacilityRequest": {
        "properties": {
          "facility_type": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "pattern": "^[a-z][a-z0-9_]*$",
            "title": "Facility Type",
            "description": "Places 設施類型代碼；只能使用已確認的類型代碼"
          },
          "origin_type": {
            "$ref": "#/components/schemas/FacilityOriginType",
            "default": "MANUAL_COORDINATE"
          },
          "origin_address": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Address"
          },
          "origin_lat": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 90.0,
                "minimum": -90.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,7}0*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Lat"
          },
          "origin_lng": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 180.0,
                "minimum": -180.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,7}0*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Lng"
          },
          "origin_reference_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Reference Id"
          },
          "confirm_lookup": {
            "type": "boolean",
            "title": "Confirm Lookup",
            "description": "必須明確設為 true 才會呼叫外部地圖服務",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "facility_type"
        ],
        "title": "NearestFacilityRequest"
      },
      "AssistantProgressResponse": {
        "properties": {
          "assistant_session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Assistant Session Id"
          },
          "current_step": {
            "type": "string",
            "title": "Current Step"
          },
          "missing_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Fields"
          },
          "missing_documents": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Documents"
          },
          "confirmed_candidate_count": {
            "type": "integer",
            "title": "Confirmed Candidate Count"
          },
          "pending_candidate_count": {
            "type": "integer",
            "title": "Pending Candidate Count"
          },
          "completed_items": {
            "type": "integer",
            "title": "Completed Items"
          },
          "total_items": {
            "type": "integer",
            "title": "Total Items"
          }
        },
        "type": "object",
        "required": [
          "assistant_session_id",
          "current_step",
          "missing_fields",
          "missing_documents",
          "confirmed_candidate_count",
          "pending_candidate_count",
          "completed_items",
          "total_items"
        ],
        "title": "AssistantProgressResponse"
      },
      "ToolExecutionResponse": {
        "properties": {
          "tool_name": {
            "type": "string",
            "title": "Tool Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Result"
          }
        },
        "type": "object",
        "required": [
          "tool_name",
          "status",
          "result"
        ],
        "title": "ToolExecutionResponse"
      },
      "KnowledgeAnswerStatus": {
        "type": "string",
        "enum": [
          "SUPPORTED",
          "EVIDENCE_ONLY",
          "CLARIFICATION_REQUIRED",
          "NO_RELEVANT_SOURCE",
          "CASE_CONTEXT_NOT_AVAILABLE"
        ],
        "title": "KnowledgeAnswerStatus"
      },
      "AssistantClaim": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "citation_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "minItems": 1,
            "title": "Citation Ids"
          }
        },
        "type": "object",
        "required": [
          "text",
          "citation_ids"
        ],
        "title": "AssistantClaim"
      },
      "AssistantCitation": {
        "properties": {
          "citation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Citation Id"
          },
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "document_title": {
            "type": "string",
            "title": "Document Title"
          },
          "document_code": {
            "type": "string",
            "title": "Document Code"
          },
          "version_no": {
            "type": "integer",
            "title": "Version No"
          },
          "effective_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective From"
          },
          "effective_to": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective To"
          },
          "page_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Start"
          },
          "page_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page End"
          },
          "section_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section Title"
          },
          "article_no": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Article No"
          },
          "quoted_text": {
            "type": "string",
            "title": "Quoted Text"
          },
          "supporting_quote": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supporting Quote"
          },
          "supported_claim": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supported Claim"
          }
        },
        "type": "object",
        "required": [
          "citation_id",
          "document_id",
          "document_title",
          "document_code",
          "version_no",
          "effective_from",
          "effective_to",
          "page_start",
          "page_end",
          "section_title",
          "article_no",
          "quoted_text"
        ],
        "title": "AssistantCitation"
      },
      "KnowledgeUnreadableSource": {
        "properties": {
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "document_title": {
            "type": "string",
            "title": "Document Title"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "document_title",
          "original_filename",
          "reason"
        ],
        "title": "KnowledgeUnreadableSource"
      },
...
      "FacilityOriginType": {
        "type": "string",
        "enum": [
          "SUBJECT_PARCEL_ENTRANCE",
          "BENCHMARK_LAND_ENTRANCE",
          "DISTRICT_REPRESENTATIVE",
          "MANUAL_COORDINATE"
        ],
        "title": "FacilityOriginType"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}